feat: move project tracking to shared goose crate
